Adjusting Hidden Mod As Speed Multiplier
As mentioned in my bug fix, this puts the change suggested live.

What this does is adjusts `_speedValue` to have a `1.18*` multiplier for the HD mod and nerfs `_aimValue` from `1.18` to `1.03` for HD. Personally I would like to adjust other things but for the sake of this being concise and clean I'll leave it like this for now. I was going to adjust the multiplier for FL but I believe that the multiplier is actually quite large as is so I am hesitant to adjust it haphazardly. 

For now this is fine I believe. I'm excited to see how this shakes up the meta! 



Note: Reading through this code really makes me want to clean up areas I think could be improved though :o