AR Fix for Aim on HD
This is a solution to the problem posed by https://github.com/ppy/osu-performance/issues/45.

I believe this to be much superior to the system currently implemented. This makes it so the maps that are problems still get nerfed (and in some cases, even harder) but there is better balance towards lower ended HD maps. Since there are people who are much better at running the actual calculations to see how this may better balance things, I will leave it up to them.